fix: Separate server and client auth code to resolve Vercel build error
- Create dev-auth-server.ts with server action for user creation
- Remove server-side imports from dev-auth.ts (fixes next/headers error)
- Use createDevUserAction server action instead of direct admin client
- Resolves 'next/headers only works in Server Component' build failure

This fixes the Vercel deployment issue where server-side code was being
imported in client-side components.
